Rename gen-ai-schema-validation to otel-schema-validator and fix package.json ordering
