+++ aws-distro-opentelemetry-node-autoinstrumentation/package.json
@@ -82,12 +82,11 @@
     "@aws-sdk/client-s3": "3.982.0",
     "@aws-sdk/client-secrets-manager": "3.982.0",
     "@aws-sdk/client-sfn": "3.982.0",
-    "@aws-sdk/client-sns": "3.982.0",
     "@aws-sdk/client-sts": "3.982.0",
+    "@aws-sdk/client-sns": "3.982.0",
     "@langchain/anthropic": "^1.3.26",
     "@langchain/aws": "^1.3.5",
     "@langchain/cohere": "^1.0.4",
-    "@langchain/core": "^1.0.0",
     "@langchain/deepseek": "^1.0.24",
     "@langchain/google-genai": "^2.1.27",
     "@langchain/groq": "^1.2.0",
@@ -102,6 +101,7 @@
     "@types/mocha": "7.0.2",
     "@types/node": "18.6.5",
     "@types/proxyquire": "^1.3.31",
+    "@langchain/core": "^1.0.0",
     "@types/sinon": "10.0.18",
     "expect": "29.2.0",
     "mocha": "7.2.0",
@@ -123,12 +123,12 @@
     "@opentelemetry/auto-instrumentations-node": "0.72.0",
     "@opentelemetry/baggage-span-processor": "0.5.0",
     "@opentelemetry/core": "2.6.1",
-    "@opentelemetry/exporter-logs-otlp-grpc": "0.214.0",
-    "@opentelemetry/exporter-logs-otlp-http": "0.214.0",
-    "@opentelemetry/exporter-logs-otlp-proto": "0.214.0",
     "@opentelemetry/exporter-metrics-otlp-grpc": "0.214.0",
     "@opentelemetry/exporter-metrics-otlp-http": "0.214.0",
     "@opentelemetry/exporter-trace-otlp-proto": "0.214.0",
+    "@opentelemetry/exporter-logs-otlp-grpc": "0.214.0",
+    "@opentelemetry/exporter-logs-otlp-http": "0.214.0",
+    "@opentelemetry/exporter-logs-otlp-proto": "0.214.0",
     "@opentelemetry/exporter-zipkin": "2.6.1",
     "@opentelemetry/id-generator-aws-xray": "2.1.0",
     "@opentelemetry/instrumentation": "0.214.0",
