+++ aws-distro-opentelemetry-node-autoinstrumentation/test/instrumentation/otel-schema-validator.ts
